Change to logic on setting flat exposure time.
Now, if the DifferentFlatExposure PV is set to "Same",
the FlatExposureTime PV will be set to equal the ExposureTime
PV.  This will help in recording the right flat exposure time
in dx files.

Also change the front end shutter logic for 7-BM to monitor the
shutter status.  This was taken from the 2-BM code.