Add a script to auto start the IOC.
This is based on a shell script APS BCDA has used in their soft IOCs
for some time.  It automatically starts IOCs in a screen session,
so one does not have to worry about the IOC dying if a terminal
window is closed.  It ensures that two instances of TomoScan are
not started on the same machine.  It also allows for easy stopping of
the IOC.